### 2019-03-11 #1307 @kyleniemeyer
@karthik this submission looks like it might be up your alley—can you edit it?
