### 2019-03-15 #1307 @FranzKrah
List of potential reviewers:
- emmanuel.paradis@ird.fr
- juancamilovillada@gmail.com
- me@nanx.me
- shaunpwilkinson@gmail.com
- frederick.boehm@gmail.com
- mdsherman@betteridiot.tech

and should I add the missing DOIs?
