### 2019-03-19 #1307 @arfon
> and should I add the missing DOIs?

@FranzKrah - yes please. The list of DOIs above ☝️should be considered as suggestions, please check them before adding them.

> Hi, I could need some assistance here! Is there anything I should do now?

I think the next steps are on @karthik here to take your reviewer suggestions and invite the reviewers to participate.
