+++ paper.bib
@@ -1,10 +1,11 @@
-@article{Muenchen2014,
+@webpage{Muenchen2014,
 	Author = {Muenchen, Robert A},
 	Date-Added = {2019-02-08 06:45:09 +0100},
-	Date-Modified = {2019-02-08 06:45:10 +0100},
-	Journal = {no. April},
+	Date-Modified = {2019-03-15 12:52:33 +0530},
 	Title = {{The popularity of data analysis software}},
-	Year = {2014}}
+	Url = {Retrieved from: http://r4stats.com/articles/popularity},
+	Year = {2019}}
+
 	
 @article{Cavender-BaresKozak2009,
 	Abstract = {The increasing availability of phylogenetic data, computing power and informatics tools has facilitated a rapid expansion of studies that apply phylogenetic data and methods to community ecology. Several key areas are reviewed in which phylogenetic information helps to resolve long-standing controversies in community ecology, challenges previous assumptions, and opens new areas of investigation. In particular, studies in phylogenetic community ecology have helped to reveal the multitude of processes driving community assembly and have demonstrated the importance of evolution in the assembly process. Phylogenetic approaches have also increased understanding of the consequences of community interactions for speciation, adaptation and extinction. Finally, phylogenetic community structure and composition holds promise for predicting ecosystem processes and impacts of global change. Major challenges to advancing these areas remain. In particular, determining the extent to which ecologically relevant traits are phylogenetically conserved or convergent, and over what temporal scale, is critical to understanding the causes of community phylogenetic structure and its evolutionary and ecosystem consequences. Harnessing phylogenetic information to understand and forecast changes in diversity and dynamics of communities is a critical step in managing and restoring the Earth's biota in a time of rapid global change.},
@@ -86,6 +87,7 @@
 	Date-Added = {2016-10-08 13:28:39 +0000},
 	Date-Modified = {2019-01-01 13:59:32 +0100},
 	Journal = {Bioinformatics},
+	Doi = {10.1093/bioinformatics/btu033},
 	Number = {9},
 	Pages = {1312--1313},
 	Publisher = {Oxford Univ Press},
@@ -98,6 +100,7 @@
 	Date-Added = {2017-02-25 18:36:06 +0000},
 	Date-Modified = {2017-02-25 18:36:13 +0000},
 	Journal = {Molecular Biology and Evolution},
+	Doi = {10.1093/molbev/msq066},
 	Number = {8},
 	Pages = {1759--1767},
 	Publisher = {SMBE},
@@ -111,6 +114,7 @@
 	Date-Added = {2017-01-17 13:40:20 +0000},
 	Date-Modified = {2017-01-17 13:40:26 +0000},
 	Journal = {Nucleic acids research},
+	Doi = {10.1093/nar/gkv318},
 	Number = {W1},
 	Pages = {W7--W14},
 	Publisher = {Oxford Univ Press},
@@ -123,6 +127,7 @@
 	Booktitle = {Pacific Symposium on Biocomputing},
 	Date-Added = {2017-03-16 14:11:37 +0000},
 	Date-Modified = {2017-03-16 14:11:39 +0000},
+	Doi = {10.1142/9789812776136_0003},
 	Pages = {15--24},
 	Title = {Local reliability measures from sets of co-optimal multiple sequence alignments.},
 	Volume = {13},
@@ -133,6 +138,7 @@
 	Date-Added = {2018-08-09 15:13:08 +0200},
 	Date-Modified = {2018-10-19 10:37:37 +0200},
 	Journal = {BMC Evolutionary Biology},
+	Doi = {10.1186/s12862-018-1229-7},
 	Number = {1},
 	Pages = {119-132},
 	Publisher = {BioMed Central},
