### 2019-03-26 #1350 @whedon
[REVIEW]: rGUIDANCE – alignment confidence score computation in R
**Submitting author:** @FranzKrah (<a href="http://orcid.org/0000-0001-7866-7508">Franz-Sebastian Krah</a>)
**Repository:** <a href="https://github.com/FranzKrah/rGUIDANCE" target ="_blank">https://github.com/FranzKrah/rGUIDANCE</a>
**Version:** 1.0

@shaunpwilkinson , please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
