### 2019-03-26 #1350 @karthik
@shaunpwilkinson Please follow instructions here and work through the checklist. Let me know if you have any questions. You can use Whedon to generate a new pdf anytime with `@whedon generate pdf` or check references with `@whedon check references`
