### 2019-03-27 #1350 @FranzKrah
I checked the proof version and I am not sure why but the Figure captions are not there...
Franz
