### 2019-03-28 #1350 @karthik
Franz: Thanks for catching that. Your figure captions render locally for me but we'll work to get this sorted out before publication.
