### 2019-04-28 #1350 @shaunpwilkinson
This is a very helpful package that fills a critical gap in the R workflow for phylogenetic analysis. The accompanying paper is clear and well written, and an example is included demonstrating the use of the package to improve the resolution of a phylogeny by factoring in column-wise alignment uncertainty.
The tutorial needs some refining and additional notes to help users through the process without encountering issues, particularly when calling the third-party executables (outlined below). The example also involves downloading sequences from NCBI, which means that a set of sequences downloaded may change from one day to the next. This caused me a few issues and required some troubleshooting. For example, the outgroup _Helvella aestivalis_ was not included in the returned sequence set when I ran the script, so the raxml calls produced an error. Also the names of the alignment caused raxml to fail with a rather enigmatic error message (it would be worth noting in the tutorial that the sequence names need to be truncated, not include special metacharacters, and all be unique). To avoid these issues I would suggest either including a pre-downloaded sequence data file with the package or use an existing sequence dataset such as the woodmouse alignment (**ape** package).

There is also a vital line missing from the example script, `run(wd = wd)` should be added before `all_clusters <- read_phylota(wd)`.

Some other minor issues are included below:

The full list of dependencies should be included in the tutorial, along with links to the various repositories, optionally some basic installation instructions for Windows, Mac and Linux, and additional annotation/notes should be included to make it clear where paths to executables need to be modified. Some tips on finding the executables within the file system may also be helpful for some users.

The latest version of the **ips** package on CRAN is v0.0.7, but the **rGUIDANCE** package depends on v0.0.11. Please either add `devtools::install_github("heibl/ips")` to the example script or include the **ips** github repo in the ‘Remotes’ field of the DESCRIPTION file.

The example script includes local directories that won’t apply for other users – e.g. `wd <- 'Documents/PhD/proj/low_priority/phylotaR/helvella'`. I would advise just using `'[path-to-directory]'` or something similar. Alternatively rOpenSci may have a standard for setting working directory in example scripts. @karthik may be able to advise on this?

Consider changing the number of threads in the example to something more manageable for non-server users. I would advise changing the `ncore` argument to 1 for the purposes of the tutorial.

`str_extract` needs the **stringr** package installed; this should be added to the example script. In this and other calls it would be safer to use the double-colon operator to specify the package, since the function called may depend on the order that the various packages were loaded. I would also specify the full method where applicable too, for example use `w <- phylogram::as.dendrogram.phylo(w)` instead of `w <- as.dendrogram(w)`.

On a final note, I found the manuscript provided a really helpful and intuitive guide to using R for phylogenetic analysis. In case of interest (and at the risk of a plug) the **aphid** package also does multiple sequence alignment within R and doesn't require any third party programs.

I hope this helps - thanks for the opportunity to review. I'm looking forward to including this package in my own workflows.

All the best,
Shaun
