### 2019-04-28 #1350 @karthik
Thanks @shaunpwilkinson!
@FranzKrah Can you please make these fixes and respond here? I opened a couple of issues on your repo.
