+++ vignettes/Phylogeny_pipeline.Rmd
@@ -11,6 +11,34 @@ vignette: >
 
 # Simple (examplary) pipeline for phylogeny inference integrating MSA confidence scores
 
+## Dependencies
+
+The R package rGUIDANCE has many dependencies. Some are external of R (R Core Team 2017), some are internal.
+
+### External
+rGUIDANCE uses third-party executables (e.g., RAxML, MAFFT). Following a list with URLs to where these programs can be downloaded. Within the following script only three are needed: MAFFT, RAxML and BLAST (dependency of another R package). For more information please refer to the guidance help page (just type ?guidance).
+
+* MAFFT: https://mafft.cbrc.jp/alignment/software/
+* CLUSTAL: http://www.clustal.org/clustal2/
+* PRANK: http://wasabiapp.org/software/prank/
+* MUSCLE: https://www.ebi.ac.uk/Tools/msa/muscle/
+* RAxML:  https://cme.h-its.org/exelixis/web/software/raxml/index.html
+* BLAST: https://blast.ncbi.nlm.nih.gov/Blast.cgi?CMD=Web&PAGE_TYPE=BlastDocs&DOC_TYPE=Download
+
+Functions which utilize any of these programs usually contain the argument "exec" (for executable). exec is the path to where the executable lies on the local computer. The functions in R then parse information towards the program and captures the results back into R. 
+
+If you have trouble where to find executables, please refer to guides such as:
+Windows: https://tiptopsecurity.com/how-to-find-the-executable-file-for-a-program/
+Mac: https://www.alvinpoh.com/how-to-find-the-location-of-an-executable-program-in-the-mac-terminal/
+
+
+### Internal
+rGUIDANCE depends on many R packages. The most important are: Rcpp, ape, ips and foreach. Rcpp allows to out-source heavy computations into C++ code via R (Eddelbuettel and Francois 2011). The score calculations involve many pairwise comparisons between alignments which is slow in R. ape is a basic R package for the "Analyses of Phylogenetics and Evolution" (Paradis, Claude, and Strimmer 2004; Paradis 2012) and was especially used for handling the alternative guide trees. ips hosts interfaces to other software, such as RAxML (ips = Interfaces to Phylogenetic Software). foreach allows easy parallelization of for loops (Microsoft and Weston 2017). 
+
+## Cautionary note
+The below example involves downloading sequences from NCBI GenBank, which means that the sequences downloaded may have changed when you run the script. Therefore we put the sequences (cluster 0; cl0.fas) as datafile in the R package as an example file. We here sketched out the pipeline described in the rGUIDANCE paper. If you only want to see how guidance works on an alignment and downstream analysis then directly jump to point 2 after installing rGUIDANCE and other packages.
+
+
 ## Install rGUIDANCE
 ```{r guidance,  echo=TRUE, eval=FALSE}
 ## Install rGUIDANCE
@@ -24,15 +52,17 @@ library(phytools)
 library(rGUIDANCE)
 ```
 
+
 ## 1. Use phylotaR to retrieve sequence clusters 
 ```{r phyloraR, echo=TRUE, eval=FALSE}
 library(phylotaR)
[…]
