+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: rGUIDANCE
 Type: Package
 Title: Sequence alignment confidence score computation in R
-Version: 1.0
+Version: 1.0.1
 Date: 2019-02-08
 Authors@R: c(person("Franz-Sebastian", "Krah", email = "f.krah@mailbox.org",
                   role = c("aut", "cre")),
