### 2019-04-30 #1350 @karthik
@FranzKrah One more change (i.e. the title). It currently has your GitHub handle in the software title too. Can you fix that as well?  Current citation:

```
Franz Sebastian Krah, & Christoph Heibl. (2019, April 30). FranzKrah/rGUIDANCE v1.0.8 (Version v1.0.8). Zenodo. http://doi.org/10.5281/zenodo.2654272
```

Otherwise the citation for the software archive will also have your Github handle rather than just the name of your software. See a [correct example](https://zenodo.org/record/2650709)
