### 2019-04-30 #1350 @FranzKrah
@karthik I changed the title but I cannot get rid of "FranzKrah/rGUIDANCE: ".
I tried already to include a .zenodo.json file in the root of my GitHub repo but this creates errors in Zenodo...
